Immutable RealmSchema and RealmObjectSchema
Before this change, Realm.getSchema() and DynamicRealm.getSchema() both
return a mutable version RealmSchema object. That would cause issue when
changing the typed Realm's schema then continue to use typed inferface
to access the Realm where the column indices might be changed already
and had not been refreshed before the transaction commited.

After this change:
Realm.getSchema() returns an immutable RealmSchema object. And all
RealmObjectSchema objects retrieved from that will be immutable as
well.
