+++ realm/realm-library/src/main/java/io/realm/DynamicRealm.java
@@ -49,12 +49,16 @@ import rx.Observable;
  */
 public class DynamicRealm extends BaseRealm {
 
+    private final RealmSchema schema;
+
     private DynamicRealm(RealmCache cache) {
         super(cache);
+        this.schema = new MutableRealmSchema(this);
     }
 
     private DynamicRealm(RealmConfiguration configuration) {
         super(configuration);
+        this.schema = new MutableRealmSchema(this);
     }
 
     /**
@@ -253,6 +257,16 @@ public class DynamicRealm extends BaseRealm {
     @Override
     public Observable<DynamicRealm> asObservable() {
         return configuration.getRxFactory().from(this);
+    }
+
+    /**
+     * Returns the mutable schema for this Realm.
+     *
+     * @return The {@link RealmSchema} for this Realm.
+     */
+    @Override
+    public RealmSchema getSchema() {
+        return schema;
     }
 
     /**
